♻️ Refactor: Separate Header with Post
